+++ aderyn_driver/src/lib.rs
@@ -6,14 +6,4 @@ pub mod driver;
 pub mod lsp_report;
 pub mod preprocess;
 
-use std::path::{Path, PathBuf};
-
 pub use aderyn_core::{ast as core_ast, detect as detection_modules, detect::detector};
-
-fn ensure_valid_root_path(root_path: &Path) -> PathBuf {
-    if !root_path.exists() {
-        eprintln!("{} does not exist!", root_path.to_string_lossy());
-        std::process::exit(1);
-    }
-    std::fs::canonicalize(root_path).unwrap()
-}
